+++ core/shared/src/test/scala/fs2/concurrent/TopicSuite.scala
@@ -36,7 +36,7 @@ class TopicSuite extends Fs2Suite {
         .covary[IO]
         .through(topic.publish)
       val subscriber =
-        topic.subscribe(Int.MaxValue).take(count + 1).fold(Vector.empty[Int]) {
+        topic.subscribe(Int.MaxValue).take(count.toLong + 1).fold(Vector.empty[Int]) {
           _ :+ _
         }
 
@@ -71,7 +71,7 @@ class TopicSuite extends Fs2Suite {
           .through(topic.publish)
         val subscriber = topic
           .subscribe(1)
-          .take(count + 1)
+          .take(count.toLong + 1)
           .flatMap(is => Stream.eval(signal.get).map(is -> _))
           .fold(Vector.empty[(Int, Int)])(_ :+ _)
